Rename hexAddress.ts to hex.ts
